requirements.txt
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
cycler==0.10.0
decorator==5.1.0
empyrical==0.5.5
idna==3.2
ipython==7.28.0
jedi==0.18.0
joblib==1.0.1
kiwisolver==1.3.2
llvmlite==0.37.0
lxml==4.6.3
matplotlib==3.4.3
matplotlib-inline==0.1.3
mysql-connector-python==8.0.26
numba==0.54.0
numpy==1.20.3
pandas==1.3.3
pandas-datareader==0.10.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plotly==5.3.1
prompt-toolkit==3.0.20
protobuf==3.18.0
ptyprocess==0.7.0
pyfolio==0.9.2
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
tenacity==8.0.1
threadpoolctl==3.0.0
traitlets==5.1.0
urllib3==1.26.7
wcwidth==0.2.5
